oc: separate test for the deploymentconfig describer
